test: fix qunit start the test too early
Now that the markdown plugin is async, the ready events are not fired quick enough for the assertions to be register before qunit start running the tests.

The solution is to have all decks resolve their promise when they finish register the tests. The qunit is then instructed to wait on these promises to be resovled before starting the test.